\name{genomes-internal}
\alias{abbrev}
\alias{bases}
\alias{extract.genomes}
\alias{[.genomes}
\alias{latin2char}
\alias{tax2id}
\alias{top}
\alias{updatelproks}
\alias{updateproks}
\alias{updateeuks}
\alias{updatevirus}
\alias{xattr}
\alias{xtags}
\alias{xvalue}
\title{Internal genomes Functions}
\description{
  Internal genomes functions.
}
\note{
  The update functions contain code to download and format genome
  tables. Others are internal.
}
\keyword{internal}
