\name{genomes-internal}
\alias{abbrev}
\alias{updatelenvs}
\alias{updateleuks}
\alias{updatelproks}
\alias{updatevirus}
\alias{latin2char}
\alias{extract.genomes}
\alias{[.genomes}

\title{Internal genomes Functions}
\description{
  Internal genomes functions.
}
\note{
  The update functions contain code to download and format genome
  tables. Others are internal.
}
\keyword{internal}
